feat(graph-layout): add WASM binding via core split
Extracts the Sugiyama-style hierarchical layout (rank assignment +
crossing reduction + coordinate assignment) into `_graph-layout-core`.
Pure compute, no deps. The napi crate is a thin shim; WASM exposes
the same `layout(spec)` / `layoutMany(specs)` single-FFI-crossing
surface via serde-wasm-bindgen.
