+++ crates/graph-layout/wasm/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "amigo-graph-layout-wasm"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+amigo-graph-layout-core = { path = "../../_graph-layout-core" }
+wasm-bindgen = { workspace = true }
+serde-wasm-bindgen = { workspace = true }
+serde = { version = "1", features = ["derive"] }
+
+[dev-dependencies]
+wasm-bindgen-test = { workspace = true }
+serde_json = "1"
+
+[package.metadata.wasm-pack.profile.release]
+wasm-opt = false
